orchestrator/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
logger==1.4
uuid6==1.0.3
